requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
python-dotenv==0.21.1
python-substack==0.1.9
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.4
